requirements.txt
streamlit>=1.32.0
openai>=1.30.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
einops>=0.7.0
